Update README.md
Student Attendance System” is developed for maintaining the attendance of the student on the daily basis in the Schools/Colleges. Here the student will be responsible to mark the attendance using provided barcode. An accurate report based on student attendance is generated here. This system will also help in evaluating the attendance eligibility criteria of a student. A report of the student’s attendance on a weekly and monthly basis is generated